feat: Added a wrapper for the zephyr driver of the bmi088
Referes to the zephyr migration

+++ src/utils/Vector3f.h
@@ -3,27 +3,27 @@
 
 struct Vector3f
 {
-    float x,y,z;
+    double x,y,z;
 
 
     Vector3f() : x(0),y(0),z(0) {}
-    Vector3f(float x, float y, float z) : x(x), y(y), z(z) {}
+    Vector3f(double x, double y, double z) : x(x), y(y), z(z) {}
 
     Vector3f operator+(const Vector3f& v) const { return {x+v.x, y+v.y, z+v.z}; }
     Vector3f operator-(const Vector3f& v) const { return {x-v.x, y-v.y, z-v.z}; }
-    Vector3f operator*(float s) const { return {x*s, y*s, z*s}; }
-    Vector3f operator/(float s) const { return {x/s, y/s, z/s}; }
+    Vector3f operator*(double s) const { return {x*s, y*s, z*s}; }
+    Vector3f operator/(double s) const { return {x/s, y/s, z/s}; }
 
      // Prodotti
-    float dot(const Vector3f& v) const { return x*v.x + y*v.y + z*v.z; }
+    double dot(const Vector3f& v) const { return x*v.x + y*v.y + z*v.z; }
     Vector3f cross(const Vector3f& v) const {
         return { y*v.z - z*v.y, z*v.x - x*v.z, x*v.y - y*v.x };
     }
 
     // Utility
-    float norm() const { return std::sqrt(x*x + y*y + z*z); }
+    double norm() const { return std::sqrt(x*x + y*y + z*z); }
     void normalize() {
-        float n = norm();
+        double n = norm();
         if (n > 1e-6f) { x /= n; y /= n; z /= n; }
     }
 };
