+++ src/utils/srimb_topics/imu_topic/imu_topic.hpp
@@ -14,7 +14,7 @@ struct ImuData {
     Vector3f filtered_acc;          
     Vector3f filtered_gyro;
 
-    int16_t temp;
+    double temp;
     uint8_t imu_id = 0;
     //Da aggiungere scoring e stato sensore e capire se mettere quaternioni qua(quindi se fare la conversione nel driver o nel imuManager)
 };
